pyproject.toml
[project]
name = "pedidos-huevo-streamlit"
version = "0.1.0"
description = "Miniapp interna para captura semanal de pedidos de huevo."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "streamlit>=1.43.0",
  "pandas>=2.2.0",
  "openpyxl>=3.1.5",
  "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pedidos_huevo"]